requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
assemblyai==0.26.0
attrs==23.2.0
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
decorator==4.4.2
Django==5.0.6
frozenlist==1.4.1
g4f==0.3.1.9
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
moviepy==1.0.3
multidict==6.0.5
numpy==1.26.4
Pillow==9.5.0
proglog==0.1.10
pycryptodome==3.20.0
pydantic==2.7.2
pydantic_core==2.18.3
python-dotenv==1.0.1
requests==2.32.3
setuptools==70.0.0
sniffio==1.3.1
sqlparse==0.5.0
srt==3.5.3
srt_equalizer==0.1.9
tqdm==4.66.4
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
websockets==12.0
yarl==1.9.4
